flask-app/requirements.txt
Flask==2.3.2
opentelemetry-api==1.19.0
opentelemetry-instrumentation-flask==0.40b0
opentelemetry-instrumentation-requests==0.40b0
prometheus_client==0.17.1
gunicorn==20.1.0
opentelemetry-semantic-conventions==0.40b0
python-json-logger==2.0.7
requests==2.31.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-api==1.19.0
opentelemetry-instrumentation-logging==0.40b0